Outline shader complete
Now allows outlines to be added to objects by giving the object 2 materials (one that is its own and the second being the outline material) and the Outline script, and assigning the same materials to its fields. Functions for activating and deactivating the outlines needs to be called elsewhere (though it holds the functionality for OnMouseOver() actions for testing purposes.